# admin: Synchronize server assignment/references to avoid data race